part 4 complete

+++ part4/controllers/bloglist.js
@@ -1,8 +1,9 @@
 const bloglistRouter = require('express').Router()
 const Blog = require('../models/blog')
 
+
 bloglistRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
@@ -16,14 +17,30 @@ bloglistRouter.get('/:id', async (request, response) => {
 })
 
 bloglistRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
-  await blog.save()
-  response.status(201).json(blog)
+  const user = request.user
+  const blog = new Blog({
+    ...request.body,
+    user: user.id
+  })
+
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
+  response.status(201).json(savedBlog)
 })
 
 bloglistRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+  if (!blog.user || blog.user.toString() === user.id.toString()) {  // allow all users to delete old data if available
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(401).json({
+      error: 'unauthorized token for deleting the blog'
+    })
+  }
 })
 
 bloglistRouter.put('/:id', async (request, response) => {
